NPCTemplates and resets load basic info. Act function.
Resets are called once when the game loads. This will be changed later.

Newline prepended to output if sitting at prompt. New line appended before prompt is displayed if output doesn't end with one already.

Only alphabet characters can be used in a name and it must be between 3 and 12 characters in length.